Use intptr_t and uintptr_t for sx and ux aliases (host pointer width)
- The sx and ux types may need to by aliased later to support RV32.
  The purpose is to use a placeholder type globally for the natural
  register width of the target. These placeholder types match the
  C pseudo-code cast notation used in meta/instructions

+++ src/riscv-types.h
@@ -22,8 +22,8 @@ namespace riscv {
 	 * Short-hand type aliases
 	 */
 
-	typedef signed long long   sx;
-	typedef unsigned long long ux;
+	typedef intptr_t           sx;
+	typedef uintptr_t          ux;
 	typedef int8_t             s8;
 	typedef uint8_t            u8;
 	typedef int16_t            s16;
